% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.basic.R
\name{germline.somatic.assoc}
\alias{germline.somatic.assoc}
\title{Germline-somatic association test}
\usage{
germline.somatic.assoc(
  y.vals,
  x.vals,
  samples,
  meta,
  firth.fallback = TRUE,
  firth.always = FALSE,
  custom.covariates = c(),
  multiPop.min.ac = 10,
  multiPop.min.freq = 0.01
)
}
\arguments{
\item{y.vals}{Numeric vector indicating presence or absence of somatic endpoint}

\item{x.vals}{Numeric vector of germline values to test}

\item{samples}{Vector of sample IDs to consider}

\item{meta}{Metadata for all samples as loaded by \link{load.patient.metadata}}

\item{firth.fallback}{Attempt to use Firth bias-reduced logistic regression when
traditional logistic regression fails to converge or dataset is quasi-separable
[default: TRUE]}

\item{firth.always}{Always use Firth regression [default: FALSE]}

\item{custom.covariates}{Character vector of columns from \code{meta} to include
as covariates in the regression model. See \code{Details}.}

\item{multiPop.min.ac}{Minimum allele count to include non-European populations}

\item{multiPop.min.freq}{Minimum variant frequency to include non-European populations}
}
\value{
Vector of test results
}
\description{
Run association test for a single germline-somatic interaction
}
\details{
By default, the following covariates will be included:
\itemize{
\item Age, sex, and an interaction term for age by sex
\item Top ten principal components \(PCs\)
\item Tumor purity
}

Any other covariates to be included must be specified in \code{custom.covariates}
}
\seealso{
\link{load.patient.metadata}
}
