% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{get.cohort.palette}
\alias{get.cohort.palette}
\title{Get Multi-Cohort Color Palette}
\usage{
get.cohort.palette(palette, cohorts)
}
\arguments{
\item{palette}{Palette of colors to be sampled}

\item{cohorts}{Character vector of cohorts to extract}
}
\description{
Extract color shades from an existing palette corresponding to specific cohort labels
}
\examples{
load.constants("colors")
get.cohort.palette(PDAC.colors, c("DFCI", "TCGA"))

}
\seealso{
\link{load.constants}
}
