% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{impute.missing.values}
\alias{impute.missing.values}
\title{Impute missing values}
\usage{
impute.missing.values(df, fill.missing = "mean", fill.columns = NULL)
}
\arguments{
\item{df}{Dataframe to impute}

\item{fill.missing}{Behavior for handling missing values. See \code{Details}. [default: "mean"]}

\item{fill.columns}{Specify in which columns missing values should be filled.}
}
\description{
Impute missing values into a dataframe
}
\details{
Recognized values for \code{fill.missing} include:
\itemize{
\item \code{"mean"} : impute missing values as the mean of all non-missing values
\item \code{"median"} : impute missing values as the median of all non-missing values
\item \code{"mode"} : impute missing values as the mode of all non-missing values
}

Note that \code{fill.missing} values of \code{"mean"} and \code{"median"} are only applicable to
columns with numeric values (e.g., age, BMI, etc.). For non-numeric columns,
the behavior of \code{fill.missing = "mode"} will always be applied.

All columns will be subjected to automated missing value imputation unless a
non-\code{NULL} value is supplied to \code{fill.columns}, in which case only the column
names passed to \code{fill.columns} will be filled.
}
