% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.meta.R
\name{ivw.meta.analysis}
\alias{ivw.meta.analysis}
\title{Meta-analyze effect sizes}
\usage{
ivw.meta.analysis(df)
}
\arguments{
\item{df}{Merged data.frame of two or more cohorts \(see \code{Details}\)}
}
\value{
data.frame
}
\description{
Meta-analyze effect sizes of all germline-somatic associations
}
\details{
\code{df} should be a data.frame generated by \link{merge.assoc.stats}
}
\seealso{
\link{load.assoc.stats.single}, \link{merge.assoc.stats}, \link{ivw.meta.single}
}
