% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.meta.R
\name{ivw.meta.single}
\alias{ivw.meta.single}
\title{Inverse-variance weighted meta-analysis}
\usage{
ivw.meta.single(x, se)
}
\arguments{
\item{x}{vector of observed effect sizes}

\item{se}{vector of observed standard errors}
}
\value{
named vector of meta-analysis effect size and standard error
}
\description{
Conduct an inverse-variance weighted meta-analysis of effect sizes
}
