% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.basic.R
\name{km.curve}
\alias{km.curve}
\title{Kaplan-Meier Curves}
\usage{
km.curve(
  surv.models,
  colors,
  group.names = NULL,
  ci.alpha = 0.15,
  legend = TRUE,
  legend.names = NULL,
  parmar = c(2, 3.5, 0.25, 4)
)
}
\arguments{
\item{surv.models}{List of one or more \code{\link[survival:summary.survfit]{survival::summary.survfit}} objects}

\item{colors}{Vector of colors for the list elements in \code{surv.models}}

\item{group.names}{(Optional) group names to assign to each list element in \code{surv.models}}

\item{ci.alpha}{Transparency value \code{alpha} for confidence interval shading [default: 0.15]}

\item{legend}{Should a legend be plotted?}

\item{legend.names}{(Optional) mapping of \code{values} to labels for legend}

\item{parmar}{Margin values passed to par()}
}
\description{
Plot Kaplan-Meier curves for one or more datasets or strata
}
\seealso{
\code{\link[survival:Surv]{survival::Surv}}, \code{\link[survival:survfit]{survival::survfit}}, \code{\link[survival:summary.survfit]{survival::summary.survfit}}
}
