% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load.patient.metadata}
\alias{load.patient.metadata}
\title{Load Patient Metadata}
\usage{
load.patient.metadata(
  file,
  fill.missing = NA,
  missing.columns = NULL,
  deduplicate = FALSE
)
}
\arguments{
\item{file}{Path to input patient metadata .tsv}

\item{fill.missing}{Behavior for handling missing values. See \code{Details}. [default: NA]}

\item{deduplicate}{Deduplicate based on patient ID / first column [default: FALSE]
See \code{Details}. [default: NULL]}

\item{fill.columns}{Specify in which columns missing values should be filled.}
}
\value{
data.frame of patient metadata
}
\description{
Load a standard patient metadata file for a single cohort
}
\details{
Recognized values for \code{fill.missing} include:
\itemize{
\item \code{NA} : do not fill any missing values
\item \code{"mean"} : impute missing values as the mean of all non-missing values
\item \code{"median"} : impute missing values as the median of all non-missing values
\item \code{"mode"} : impute missing values as the mode of all non-missing values
}

Note that \code{fill.missing} values of \code{"mean"} and \code{"median"} are only applicable to
columns with numeric values (e.g., age, BMI, etc.). For non-numeric columns,
the behavior of \code{fill.missing = "mode"} will always be applied unless \code{fill.missing}
is set to \code{NA} (the default\).

All columns will be subjected to automated missing value imputation unless a
non-\code{NULL} value is supplied to \code{fill.columns}, in which case only the column
names passed to \code{fill.columns}will be filled.
}
