% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load.variant.sets}
\alias{load.variant.sets}
\title{Load variant set membership}
\usage{
load.variant.sets(file)
}
\arguments{
\item{file}{Two-column .tsv mapping variant set IDs \(first column\) to
comma-delimited list of constituent variant IDs \(second column\)}
}
\value{
data.frame
}
\description{
Load a table linking variant sets to their constituent variant IDs
}
