% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.meta.R
\name{merge.assoc.stats}
\alias{merge.assoc.stats}
\title{Merge association stats}
\usage{
\method{merge}{assoc.stats}(stats.list)
}
\arguments{
\item{stats.list}{List of single-cohort association statistics
as loaded using \link{load.assoc.stats.single}}
}
\value{
data.frame
}
\description{
Merge association statistics across multiple cohorts
}
\seealso{
\link{load.assoc.stats.single}
}
