% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{parse.variant.set}
\alias{parse.variant.set}
\title{Parse Variant Set Members}
\usage{
parse.variant.set(vid.str)
}
\arguments{
\item{vid.str}{Character object to be parsed}
}
\value{
List of character vectors
}
\description{
Parse a string of variant IDs that comprise a variant set
}
\details{
Variant IDs should be separated by commas.

Pipe delimiters will be interpreted as multiple independent sets of variant IDs
\(as is the case for co-mutation analyses, etc.\)
}
