% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.basic.R
\name{scaled.bars}
\alias{scaled.bars}
\title{Scale-Aware Barplot Cluster}
\usage{
scaled.bars(
  values,
  colors,
  group.names = NULL,
  sep.wex = 0.05,
  title = NULL,
  legend = TRUE,
  legend.names = NULL,
  parmar = c(1, 2, 1.25, 4.5)
)
}
\arguments{
\item{values}{List of character vectors of values to plot}

\item{colors}{Named vector of colors to map to \code{values}}

\item{group.names}{(Optional) group names to assign to each list element in \code{values}}

\item{title}{(Optional) title to be printed in top-right corner}

\item{legend}{Should a legend be plotted?}

\item{legend.names}{(Optional) mapping of \code{values} to labels for legend}

\item{parmar}{Margin values passed to par()}

\item{sep.cex}{Relative width scalar for whitespace on the X-axis
between groups [default: 0.05]}
}
\description{
Plot a set of stacked barplots scaled proportional to set size
}
\details{
If \code{values} is supplied as a named list, those names will be used as
\code{group.names} unless \code{group.names} is explicitly specified. Otherwise, \code{group.names}
will be set to the ordinal value of each group in \code{values}.
}
