% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.basic.R
\name{yaxis.legend}
\alias{yaxis.legend}
\title{Legend-Axis Hybrid}
\usage{
yaxis.legend(
  legend.names,
  x,
  y.positions,
  sep.wex,
  min.label.spacing = 0.1,
  lower.limit = -Inf,
  upper.limit = Inf,
  colors = NULL,
  lwd = 3
)
}
\arguments{
\item{legend.names}{Labels to be printed in legend}

\item{x}{Where the legend will connect to the rest of the plot (in X-axis units)}

\item{y.positions}{Where should the legend labels be placed (in Y-axis units)}

\item{sep.wex}{Width expansion term for text relative to \code{x}}

\item{min.label.spacing}{Minimum distance between any two labels (in Y-axis units) [default: 0.1]}

\item{lower.limit}{No label will be placed below this value on the Y-axis [default: no limit]}

\item{upper.limit}{No label will be placed above this value on the Y-axis [default: no limit]}

\item{colors}{Line colors connecting labels to plot body [default: all black]}

\item{lwd}{Width of line connecting labels to plot body [default: 3]}
}
\description{
Add a legend to the right Y-axis with lines connecting legend labels to
specified Y positions
}
