% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load.ad.matrix}
\alias{load.ad.matrix}
\title{Load an allele dosage matrix}
\usage{
load.ad.matrix(
  file,
  sort.samples = TRUE,
  sample.subset = NULL,
  sort.variants = TRUE,
  variant.subset = NULL,
  drop.ref.records = TRUE
)
}
\arguments{
\item{file}{Tab-delimited matrix of allele dosages for samples \(columns\)
by variant IDs \(rows\)}

\item{sort.samples}{Sort samples axis [default: TRUE]}

\item{sample.subset}{Vector of sample IDs to subset. If not provided, samples
will not be subset. [default: no subsetting]}

\item{sort.variants}{Sort variant axis [default: TRUE]}

\item{variant.subset}{Vector of variant IDs to subset. If not provided, variants
will not be subset. [default: no subsetting]}

\item{drop.ref.records}{Drop variants from matrix with no observed carriers
[default: TRUE]}
}
\value{
data.frame
}
\description{
Reads an allele dosage matrix into memory with options for subsetting & sorting
}
