% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.basic.R
\name{query.ad.matrix}
\alias{query.ad.matrix}
\title{Query allele dosage matrix}
\usage{
query.ad.matrix(ad, vids, action = "verbose")
}
\arguments{
\item{ad}{One or more allele dosage matrix(es) as imported by
\link{load.ad.matrix}. See \code{Details}}

\item{vids}{Variant IDs to query}

\item{action}{Action to apply to query rows; see \code{Details}}
}
\value{
numeric vector or data.frame, depending on \code{action}
}
\description{
Subsets and optionally compresses an allele dosage matrix
}
\details{
The \code{ad} argument accepts either a single data.frame or a list
of data.frames. If a list is provided, each matrix will be queried individually
and their results will be concatenated.

Recognized values for \code{action} include:
\itemize{
\item \code{"verbose"} : return the full query matrix [default]
\item \code{"any"} : return numeric indicator if any of the query rows are non-zero
for each sample
\item \code{"sum"} : return the sum of allele dosages for all query rows per sample
}
}
